aether-system/backend/requirements.txt
# AETHER Backend Dependencies
# Production-ready stack

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Core
torch==2.1.0+cpu
transformers==4.35.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# Storage
chromadb==0.4.17
sqlalchemy==2.0.23

# Image Processing
Pillow==10.1.0

# Utilities
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
